Combine taup.arrival with taup.helper_classes.
There's really not much to the taup.arrival file.
